(fixes travis CI problems) correct identation, redistribute functions from mock.build to mock.spectra
